# REST-api/openapi.et.yaml
openapi: 3.0.0
info:
  title: Google Forms klooni API
  version: "1.0.0"
  description: >
    Google Forms sarnase teenuse OpenAPI näidisspetsifikatsioon.
    Kasutab ressursside nimedena üksikuid mitmuse nimisõnu ja /sessions autentimiseks.
servers:
  - url: /
    description: Tootmisserver
...
paths:
...
  /forms/{formId}/questions:
    get:
      tags:
        - questions
      summary: Kuva kõik konkreetse vormi küsimused
      operationId: listQuestions
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vormi küsimuste nimekiri
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
        "404":
          description: Vormi ei leitud
        "401":
          description: Autoriseerimata
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]

    post:
      tags:
        - questions
      summary: Loo uus küsimus konkreetsele vormile
      operationId: createQuestion
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionCreate"
      responses:
        "201":
          description: Küsimus on loodud
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Vormi ei leitud
        "401":
          description: Autoriseerimata
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]
  /forms/{formId}/questions/{questionId}:
    get:
      tags:
        - questions
      summary: Kuva konkreetse küsimuse üksikasjad
      operationId: getQuestion
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Küsimuse üksikasjad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Küsimust ei leitud
        "401":
          description: Autoriseerimata
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]

    patch:
      tags:
        - questions
      summary: Uuenda olemasoleva küsimuse osa
      operationId: updateQuestion
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Küsimus on uuendatud
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Küsimust ei leitud
        "401":
          description: Autoriseerimata
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]

    delete:
      tags:
        - questions
      summary: Kustuta küsimus
      operationId: deleteQuestion
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Küsimus on kustutatud
        "404":
          description: Küsimust ei leitud
        "401":
          description: Autoriseerimata
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]
  /forms/{formId}/responses:
    get:
      tags:
        - responses
      summary: Kuva kõik konkreetse vormi vastused
      operationId: listResponses
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vormi vastuste nimekiri
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Response"
        "404":
          description: Vormi ei leitud
        "401":
          description: Autoriseerimata
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]

    post:
      tags:
        - responses
      summary: Loo vastus konkreetsele vormile
      operationId: createResponse
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResponseCreate"
      responses:
        "201":
          description: Vastus on loodud
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "404":
          description: Vormi ei leitud
        "401":
          description: Autoriseerimata
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]
  /forms/{formId}/responses/{responseId}:
    get:
      tags:
        - responses
      summary: Kuva konkreetse vastuse üksikasjad
      operationId: getResponse
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
        - name: responseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vastuse üksikasjad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "404":
          description: Vastust ei leitud
        "401":
          description: Autoriseerimata
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]

    patch:
      tags:
        - responses
      summary: Uuenda olemasoleva vastuse osa
      operationId: updateResponse
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
        - name: responseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResponseUpdate"
      responses:
        "200":
          description: Vastus on uuendatud
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "404":
          description: Vastust ei leitud
        "401":
          description: Autoriseerimata
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]

    delete:
      tags:
        - responses
      summary: Kustuta vastus
      operationId: deleteResponse
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
        - name: responseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Vastus on kustutatud
        "404":
          description: Vastust ei leitud
        "401":
          description: Autoriseerimata
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Vigane päring - vigased sisendparameetrid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Serveri sisemine viga
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - message
...
    QuestionCreate:
      type: object
      properties:
        text:
          type: string
        type:
          type: string
          enum: [shorttext, paragraph, multiplechoice, checkbox, dropdown]
        options:
          type: array
          items:
            type: string
      required:
        - text
        - type
    QuestionUpdate:
      type: object
      properties:
        text:
          type: string
        type:
          type: string
          enum: [shorttext, paragraph, multiplechoice, checkbox, dropdown]
    Question:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        type:
          type: string
          enum: [shorttext, paragraph, multiplechoice, checkbox, dropdown]
        options:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - text
        - type
        - createdAt
        - updatedAt
    ResponseCreate:
      type: object
      properties:
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: string
              answer:
                type: string
            required:
              - questionId
              - answer
      required:
        - answers
    ResponseUpdate:
      type: object
      properties:
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: string
              answer:
                type: string
    Response:
      type: object
      properties:
        id:
          type: string
        formId:
          type: string
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: string
              answer:
                type: string
            required:
              - questionId
              - answer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - formId
        - answers
        - createdAt
        - updatedAt
...